Entrega proyecto IKernell

+++ ikernell-frontend/src/pages/ProyectosPage.tsx
@@ -2,6 +2,9 @@ import { useEffect, useState } from "react";
 import ProjectCard from "../app/components/dashboard/ProjectCard";
 import ProyectoModal from "../app/components/dashboard/ProyectoModal";
 import { Button } from "../app/components/ui/button";
+import { obtenerIdUsuario } from "../utils/auth";
+import { toast } from "sonner";
+import { obtenerRolUsuario } from "../utils/auth";
 
 import {
   obtenerProyectos,
@@ -24,9 +27,12 @@ export default function ProyectosPage() {
     estadoProyecto: true,
 
     lider: {
-      idUsuario: 2,
+      idUsuario: obtenerIdUsuario(),
     },
   });
+
+  const rol = obtenerRolUsuario();
+
   const [proyectoEditando, setProyectoEditando] = useState<Proyecto | null>(
     null,
   );
@@ -49,19 +55,129 @@ export default function ProyectosPage() {
 
   async function guardarProyecto() {
     try {
+
+      if (!validarProyecto()) {
+        return;
+      }
+
+      function validarProyecto() {
+
+        if (!nuevoProyecto.nombreProyecto.trim()) {
+          toast.error("Ingrese un nombre");
+          return false;
+        }
+
+        if (nuevoProyecto.nombreProyecto.length < 5) {
+          toast.error("Nombre demasiado corto");
+          return false;
+        }
+
+        if (!nuevoProyecto.descripcionProyecto.trim()) {
+          toast.error("Ingrese una descripción");
+          return false;
+        }
+
+        if (
+          nuevoProyecto.descripcionProyecto.length < 10
+        ) {
+          toast.error(
+            "La descripción debe tener mínimo 10 caracteres"
+          );
+          return false;
+        }
+
+        if (!nuevoProyecto.fechaInicioProyecto) {
+          toast.error("Seleccione fecha inicio");
+          return false;
+        }
+
+        if (!nuevoProyecto.fechaFinProyecto) {
+          toast.error("Seleccione fecha fin");
+          return false;
+        }
+
+        if (
+          nuevoProyecto.fechaFinProyecto <
+          nuevoProyecto.fechaInicioProyecto
+        ) {
+          toast.error(
+            "La fecha final no puede ser menor que la inicial"
+          );
+          return false;
+        }
+
+        return true;
+      }
+
       const proyectoCreado = await crearProyecto(nuevoProyecto);
 
       setProyectos([...proyectos, proyectoCreado]);
 
+      toast.success("Proyecto creado correctamente");
+
       setMostrarFormulario(false);
     } catch (error) {
       console.error(error);
+      toast.error("No se pudo crear el proyecto");
     }
   }
 
   async function guardarCambios() {
     if (!proyectoEditando) {
       return;
+    }
+
+    if (!validarProyecto()) {
+      return;
+    }
+
+    function validarProyecto() {
+
+      if (!nuevoProyecto.nombreProyecto.trim()) {
+        toast.error("Ingrese un nombre");
+        return false;
+      }
+
+      if (nuevoProyecto.nombreProyecto.length < 5) {
+        toast.error("Nombre demasiado corto");
+        return false;
+      }
+
+      if (!nuevoProyecto.descripcionProyecto.trim()) {
+        toast.error("Ingrese una descripción");
+        return false;
+      }
+
+      if (
+        nuevoProyecto.descripcionProyecto.length < 10
+      ) {
+        toast.error(
+          "La descripción debe tener mínimo 10 caracteres"
+        );
+        return false;
+      }
+
+      if (!nuevoProyecto.fechaInicioProyecto) {
+        toast.error("Seleccione fecha inicio");
+        return false;
+      }
+
+      if (!nuevoProyecto.fechaFinProyecto) {
+        toast.error("Seleccione fecha fin");
+        return false;
+      }
+
+      if (
+        nuevoProyecto.fechaFinProyecto <
+        nuevoProyecto.fechaInicioProyecto
+      ) {
+        toast.error(
+          "La fecha final no puede ser menor que la inicial"
+        );
+        return false;
+      }
+
+      return true;
     }
 
     try {
@@ -78,21 +194,45 @@ export default function ProyectosPage() {
         ),
       );
 
+      toast.success("Proyecto actualizado correctamente");
+
       setProyectoEditando(null);
 
       setMostrarFormulario(false);
     } catch (error) {
       console.error(error);
+      toast.error("No se pudo actualizar el proyecto");
     }
   }
 
   async function borrarProyecto(id: number) {
+    const confirmar = confirm(
+      "¿Está seguro de eliminar este proyecto?"
+    );
+
+    if (!confirmar) {
+      return;
+    }
+
     try {
       await eliminarProyecto(id);
 
-      setProyectos(proyectos.filter((p) => p.idProyecto !== id));
+      setProyectos(
+        proyectos.filter(
+          (p) => p.idProyecto !== id
+        )
+      );
+
+      toast.success(
+        "Proyecto eliminado correctamente"
+      );
+
     } catch (error) {
       console.error(error);
+
+      toast.error(
+        "No se pudo eliminar el proyecto"
+      );
     }
   }
 
@@ -103,48 +243,51 @@ export default function ProyectosPage() {
   return (
     <div style={{ padding: "20px" }}>
       <h1>Proyectos</h1>
+      {rol === "Coordinador" && (
+        <Button
+          onClick={() => {
 
-      <Button
-        onClick={() => {
+            setProyectoEditando(null);
 
-          setProyectoEditando(null);
+            setNuevoProyecto({
+              codProyecto: "",
+              nombreProyecto: "",
+              descripcionProyecto: "",
+              fechaInicioProyecto: "",
+              fechaFinProyecto: "",
+              estadoProyecto: true,
 
-          setNuevoProyecto({
-            codProyecto: "",
-            nombreProyecto: "",
-            descripcionProyecto: "",
-            fechaInicioProyecto: "",
-            fechaFinProyecto: "",
-            estadoProyecto: true,
+              lider: {
+                idUsuario: obtenerIdUsuario(),
+              },
+            });
 
-            lider: {
-              idUsuario: 2,
-            },
-          });
+            setMostrarFormulario(true);
+          }}
+        >
+          Nuevo Proyecto
+        </Button>
+      )}
 
-          setMostrarFormulario(true);
-        }}
-      >
-        Nuevo Proyecto
-      </Button>
-
-      <ProyectoModal
-        open={mostrarFormulario}
-        onClose={() => {
-          setMostrarFormulario(false);
-          setProyectoEditando(null);
-        }}
-        proyecto={nuevoProyecto}
-        setProyecto={setNuevoProyecto}
-        onGuardar={
-          proyectoEditando
-            ? guardarCambios
-            : guardarProyecto
-        }
-        editando={
-          proyectoEditando !== null
-        }
-      />
+      {rol !== "Desarrollador" && (
+        <ProyectoModal
+          open={mostrarFormulario}
+          onClose={() => {
+            setMostrarFormulario(false);
+            setProyectoEditando(null);
+          }}
+          proyecto={nuevoProyecto}
+          setProyecto={setNuevoProyecto}
+          onGuardar={
+            proyectoEditando
+              ? guardarCambios
+              : guardarProyecto
+          }
+          editando={
+            proyectoEditando !== null
+          }
+        />
+      )}
 
       <div
         style={{
@@ -156,12 +299,17 @@ export default function ProyectosPage() {
         }}
       >
         {proyectos.map((proyecto) => (
+
           <ProjectCard
             key={proyecto.idProyecto}
             id={proyecto.idProyecto}
             codigo={proyecto.codProyecto}
             nombre={proyecto.nombreProyecto}
-            lider={`${proyecto.lider.nombre} ${proyecto.lider.apellido}`}
+            lider={
+              proyecto.lider
+                ? `${proyecto.lider.nombre} ${proyecto.lider.apellido}`
+                : "Sin líder"
+            }
             activo={proyecto.estadoProyecto}
             descripcion={proyecto.descripcionProyecto}
